+++ app/client/src/routes.jsx
@@ -1,13 +1,18 @@
-import HomePage from './features/home/HomePage.jsx';
 import ForbiddenPage from './features/errors/ForbiddenPage.jsx';
 import NotFoundPage from './features/errors/NotFoundPage.jsx';
 import ServerErrorPage from './features/errors/ServerErrorPage.jsx';
 import GenericErrorPage from './features/errors/GenericErrorPage.jsx';
+import HomePage from './features/home/HomePage.jsx';
+import ThemePage from './features/theme/ThemePage.jsx';
 
 const routes = [
   {
     path: '/',
     element: <HomePage />,
+  },
+  {
+    path: '/theme',
+    element: <ThemePage />,
   },
   {
     path: '/403',